requirements.txt
mysql-connector-python
python-dotenv
cryptography
cryptography
cryptography
